- Added FoodNeuralNetwork class - Loaded in Food101 dataset - Create FoodNeuralNetwork model - Trained, Tested, and Saved model

+++ main.py
@@ -4,10 +4,34 @@ import torch
 from torch import nn
 from torch.utils.data import DataLoader
 from torchvision import datasets
-from torchvision.transforms import Compose, ToTensor, Resize, CenterCrop, Normalize
+from torchvision.transforms import Compose, ToTensor, Resize, RandomCrop, Normalize, RandomHorizontalFlip
 import numpy
 from FoodNeuralNetwork import FoodNeuralNetwork
+from FoodConvNeuralNetwork import FoodConvNeuralNetwork
+import os
 
+
+foodLabels = [
+    "Apple pie", "Baby back ribs", "Baklava", "Beef carpaccio", "Beef tartare", "Beet salad",
+    "Beignets", "Bibimbap", "Bread pudding", "Breakfast burrito", "Bruschetta", "Caesar salad",
+    "Cannoli", "Caprese salad", "Carrot cake", "Ceviche", "Cheesecake", "Cheese plate",
+    "Chicken curry", "Chicken quesadilla", "Chicken wings", "Chocolate cake", "Chocolate mousse",
+    "Churros", "Clam chowder", "Club sandwich", "Crab cakes", "Creme brulee", "Croque madame",
+    "Cup cakes", "Deviled eggs", "Donuts", "Dumplings", "Edamame", "Eggs benedict", "Escargots",
+    "Falafel", "Filet mignon", "Fish and chips", "Foie gras", "French fries", "French onion soup",
+    "French toast", "Fried calamari", "Fried rice", "Frozen yogurt", "Garlic bread", "Gnocchi",
+    "Greek salad", "Grilled cheese sandwich", "Grilled salmon", "Guacamole", "Gyoza", "Hamburger",
+    "Hot and sour soup", "Hot dog", "Huevos rancheros", "Hummus", "Ice cream", "Lasagna",
+    "Lobster bisque", "Lobster roll sandwich", "Macaroni and cheese", "Macarons", "Miso soup",
+    "Mussels", "Nachos", "Omelette", "Onion rings", "Oysters", "Pad thai", "Paella", "Pancakes",
+    "Panna cotta", "Peking duck", "Pho", "Pizza", "Pork chop", "Poutine", "Prime rib",
+    "Pulled pork sandwich", "Ramen", "Ravioli", "Red velvet cake", "Risotto", "Samosa",
+    "Sashimi", "Scallops", "Seaweed salad", "Shrimp and grits", "Spaghetti bolognese",
+    "Spaghetti carbonara", "Spring rolls", "Steak", "Strawberry shortcake", "Sushi", "Tacos",
+    "Takoyaki", "Tiramisu", "Tuna tartare", "Waffles"
+]
+
+foodLabelsDict = {i: food for i, food in enumerate(foodLabels)}
 
 def validColor(value): #returns if a string is a valid color representation (a number between 0 and 255)
     return value.isnumeric() and int(value) >= 0 and int(value) <= 255
@@ -16,12 +40,22 @@ def between(a,b,c): #return if a is between b and c
     return a >= b and a <= c
 
 def eraser():
-    color[0] = 200
-    color[1] = 200
-    color[2] = 200
+    global color
+    global usingEraser
+    global pickingColor
+    color = backgroundColor.copy()
+    usingEraser = True
+    pickingColor = False
+
+def colorPicker():
+    global pickingColor
+    global usingEraser
+    pickingColor = True
+    usingEraser = False
 
 def setColor():
-    #if redInput.get_text().isnumeric() and greenInput.get_text().isnumeric() and blueInput.get_text().isnumeric() and between(int(redInput.get_text()), 0, 255) and between(int(greenInput.get_text()), 0, 255) and between(int(blueInput.get_text()), 0, 255):
+    global usingEraser
+    global pickingColor
     if validColor(redInput.get_text()) and validColor(greenInput.get_text()) and validColor(blueInput.get_text()):
             color[0] = int(redInput.get_text())
             color[1] = int(greenInput.get_text())
@@ -31,22 +65,31 @@ def setColor():
             greenInput.clear()
             blueInput.clear()
 
+            usingEraser = False
+            pickingColor = False
+
 def confirm(): #confirms the current drawing
     print("Confirm function called")
 
+    canvas_array = numpy.array(canvas)  # This creates a (height, width, 3) NumPy array
+    drawing = torch.tensor(canvas_array,dtype=torch.float32) # Convert current drawing to tensor
+    drawing = drawing.permute(2, 0, 1)  # Changes from (height, width, 3) to (3, height, width)
+
+    resize_transform = Resize((224, 224)) # Define resize transformation
+    drawing_resized = resize_transform(drawing.unsqueeze(0))  # Add batch dimension and resize, as resize expects to know how many images are being inputted
+
     with torch.no_grad(): # as we are only using the model, we don't need to back-propagate
-        drawing = torch.tensor(canvas) # convert cavnas (current drawing) to tensor
-        rawPrediction = model(drawing) # predict label based on drawing
+        rawPrediction = model(drawing_resized) # predict label based on drawing
         predictionProbabilities = nn.Softmax(dim=1)(rawPrediction) # gets the probability for each class
         prediction = predictionProbabilities.argmax(1) # pick the highest probability class as the final guess
 
-        print(f"Food: {prediction}") # displays prediction
+        print(f"Food: {foodLabelsDict[prediction.item()]}") # displays prediction
 
 
 def reset(): #resets the canvas
     for i in range(numberOfPixels):
         for j in range(numberOfPixels):
-            canvas[i][j] = (200,200,200)
+            canvas[i][j] = backgroundColor.copy()
 
 
 def drawWindow(window):
@@ -57,8 +100,6 @@ def drawWindow(window):
 
     pygame.draw.rect(window, (200, 200, 200), (windowSideLength + 37.5, 50, 75, 50)) #draws reset button
     pygame.draw.rect(window, (200, 200, 200), (windowSideLength + 37.5, 150, 75, 50))  #draws confirm button
-    pygame.draw.rect(window, (200, 200, 200), (windowSideLength + 37.5, 150, 75, 50))  # draws eraser button
-
     pygame.draw.rect(window, (200, 200, 200), (windowSideLength + 37.5, 425, 75, 30))  # draws set color button
 
     #draws pixels
@@ -98,13 +139,16 @@ def drawWindow(window):
     window.blit(greenButtontText, greenButtontTextRect)
     window.blit(blueButtontText, blueButtontTextRect)
 
-    window.blit(eraserIcon, (windowSideLength + 75, 225))
+    window.blit(eraserIcon, (windowSideLength + 80, 225))
+    window.blit(colorPickerIcon, (windowSideLength + 37.5, 225))
+
 
 # Initialize ML stuff
 
 transform = Compose([
     Resize(256),                # Resize smaller edge to 256 while preserving aspect ratio
-    CenterCrop(224),            # Crop to 224x224
+    RandomCrop(224),            # Crop to 224x224
+    RandomHorizontalFlip(),     # add a random flip to prevent overfitting
     ToTensor(),                 # Convert image to tensor
     Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])  # Normalizing to ImageNet stats
 ])
@@ -114,7 +158,7 @@ training_data = datasets.Food101(
     root="data",
     split="train",
     download=False,
-    transform=transform,
+    transform=transform
 )
 
 # Download test data from open datasets.
@@ -125,21 +169,23 @@ test_data = datasets.Food101(
     transform=transform,
 )
 
+
 # Hyperparameters
 batch_size = 64
-learning_rate = 0.001
-epochs = 5
+learning_rate = 0.01
+epochs = 7
 
 # Create data loaders (load training and testing data).
-train_dataloader = DataLoader(training_data, batch_size=batch_size)
-test_dataloader = DataLoader(test_data, batch_size=batch_size)
+train_dataloader = DataLoader(training_data, batch_size=batch_size, shuffle=True)
+test_dataloader = DataLoader(test_data, batch_size=batch_size, shuffle=False)
 
 
 # Define device on which to run model
 device = "cpu"
 
 # Create Model
-model = FoodNeuralNetwork().to(device)
+#model = FoodNeuralNetwork().to(device)
+model = FoodConvNeuralNetwork().to(device)
 
 
 # Define Loss Function and Optimizer
@@ -183,32 +229,46 @@ def test(dataloader, model, loss_fn):
     print(f"Test Error: \n Accuracy: {(100*correct):>0.1f}%, Avg loss: {test_loss:>8f} \n")
 
 
-# Train and Evaluate Model
 
-for t in range(epochs):
-    print(f"Epoch {t+1}\n-------------------------------")
-    train(train_dataloader, model, lossFunction, optimizer)
-    test(test_dataloader, model, lossFunction)
-print("Done!")
 
-# Save Model
-torch.save(model.state_dict(), "model.pth")
-print("Saved PyTorch Model State to model.pth")
+# Load Model (if already exists)
+if os.path.exists("model.pth"):
+    model.load_state_dict(torch.load("model.pth", weights_only=True))
+else:
+    # Train and Evaluate Model
+    for t in range(epochs):
+        print(f"Epoch {t+1}\n-------------------------------")
+        train(train_dataloader, model, lossFunction, optimizer)
+        test(test_dataloader, model, lossFunction)
+    print("Done!")
+
+    # Save Model
+    torch.save(model.state_dict(), "model.pth")
+    print("Saved PyTorch Model State to model.pth")
 
 
 
 color = [0,0,0]
+backgroundColor = [200, 200, 200]
 windowSideLength = 500 #window is a square so this is both length and width
 sidePanelWidth = 150
 numberOfPixels = 25 #the number of pixels in each dimension (ex. 28 means the screen is 28x28 pixels)
 sizeOfPixel = windowSideLength / numberOfPixels
 
-canvas = [[(200,200,200)] * numberOfPixels for _ in range(numberOfPixels)] #create array to store drawing
+canvas = [[backgroundColor.copy()] * numberOfPixels for _ in range(numberOfPixels)] #create array to store drawing
 
 pygame.init()
 window = pygame.display.set_mode((windowSideLength + sidePanelWidth, windowSideLength))
 pygame.display.set_caption("Whats That Drawing")
 eraserIcon = pygame.image.load("eraserIcon.png")
+colorPickerIcon = pygame.image.load("colorPickerIcon.png")
+cursor = pygame.image.load("mouseCursor.png")
+colorPickerCursor = pygame.image.load("colorPickerCursor.png")
+eraserCursor = pygame.image.load("eraserCursor.png")
+
+pygame.mouse.set_visible(False)
+pickingColor = False
+usingEraser = False
 
 # draws RGB inputs
 manager = pygame_gui.UIManager((windowSideLength + sidePanelWidth, windowSideLength))
@@ -226,13 +286,12 @@ mouseDrawing = False
 running = True
 while running:
 
+    mx, my = pygame.mouse.get_pos()
     for event in pygame.event.get():
         if event.type == pygame.QUIT: #when window is closed
             running = False
 
         if event.type == pygame.MOUSEBUTTONDOWN:  # when user clicks
-            mx, my = pygame.mouse.get_pos()
-
             if mx > windowSideLength + 37.5 and mx < windowSideLength + 112.5 and my > 50 and my < 100: # reset button pressed
                 reset()
 
@@ -242,14 +301,20 @@ while running:
             if mx > windowSideLength + 37.5 and mx < windowSideLength + 112.5 and my > 425 and my < 455: #set color button pressed
                 setColor()
 
-            if mx > windowSideLength + 75 and mx < windowSideLength + 105 and my > 225 and my < 255: #eraser icon is pressed
+            if mx > windowSideLength + 80 and mx < windowSideLength + 110 and my > 225 and my < 255: #eraser icon is pressed
                 eraser()
 
-            if mx < windowSideLength: #if click is on the canvas
-                mouseDrawing = True
+            if mx > windowSideLength + 37.5 and mx < windowSideLength + 67.5 and my > 225 and my < 255: #colorPicker icon is pressed
+                colorPicker()
+
 
             if mx > 0 and mx < 500 and my > 0 and my < 500: #check if mouse is within canvas
-                canvas[int(mx / sizeOfPixel)][int(my / sizeOfPixel)] = color.copy()
+                if pickingColor: #set the color to that of the pixel if using colorPicker tool
+                    color = canvas[int(mx / sizeOfPixel)][int(my / sizeOfPixel)].copy()
+                    pickingColor = False
+                else: #otherwise draw the pixel
+                    canvas[int(mx / sizeOfPixel)][int(my / sizeOfPixel)] = color.copy()
+                    mouseDrawing = True
 
 
         if mouseDrawing and event.type == pygame.MOUSEMOTION: #if the user is drawing, and moves the mouse
@@ -266,6 +331,13 @@ while running:
         manager.process_events(event)
 
     drawWindow(window)
+
+    if pickingColor:
+        window.blit(colorPickerCursor, (mx, my))  # draws colorPickerCursor icon
+    elif usingEraser:
+        window.blit(eraserCursor, (mx, my))  # draws eraserCursor icon
+    else:
+        window.blit(cursor, (mx, my))  # draws regular mouse icon
 
 
     # Update display
